Remove code to display Hecke charpolys on newspace pages. Refactor to reflect that we now only display these for newform pages. Newspace tests removed (pass locally).
