+++ lmfdb/classical_modular_forms/templates/cmf_full_gamma1_space.html
@@ -179,8 +179,4 @@ function show_qexp(qstyle) {
 </div>
 {% endif %}
 
-<h2>{{KNOWL('cmf.heckecharpolys', title='Hecke characteristic polynomials')}}</h2>
-
-{{ space.display_hecke_char_polys() | safe }}
-
 {% endblock %}
